+++ test/dummy/config/initializers/resque.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+require "mock_redis"
+
+Resque.redis = MockRedis.new